x501: move `Time`/`Validity` from `x509` (#388)
These types were originally defined in X.501. They are also used more
broadly than just in X.509.
