+++ x501/Cargo.toml
@@ -17,6 +17,9 @@ hex-literal = "0.3"
 [dependencies]
 der = { version = "=0.6.0-pre.1", features = ["derive", "alloc", "oid"], path = "../der" }
 
+[features]
+std = ["der/std"]
+
 [package.metadata.docs.rs]
 all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
