+++ x501/src/lib.rs
@@ -1,4 +1,4 @@
-#![no_std]
+#![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![doc = include_str!("../README.md")]
 #![doc(
@@ -13,3 +13,4 @@ extern crate alloc;
 
 pub mod attr;
 pub mod name;
+pub mod time;
